pyproject.toml
[project]
name = "codeocean-mcp-server"
version = "0.1.0"
authors = [
  { name="Code Ocean", email="dev@codeocean.com" },
]
description = "Code Ocean MCP Server"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "codeocean>=0.5.0",
    "fastmcp>=2.5.1",
]
license = "MIT"

[dependency-groups]
dev = [
    "hatch>=1.14.1",
    "ruff>=0.11.11",
]

[project.urls]
Homepage = "https://github.com/codeocean/codeocean-mcp-server"
Issues = "https://github.com/codeocean/codeocean-mcp-server/issues"
Changelog = "https://github.com/codeocean/codeocean-mcp-server/blob/main/CHANGELOG.md"

[project.scripts]
codeocean-mcp-server = "codeocean_mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.circleci",
]

[tool.hatch.build.targets.wheel]
packages = ["src/codeocean_mcp_server"]

[tool.hatch.envs.default.scripts]
test = "python -m unittest -v"

[[tool.hatch.envs.test.matrix]]
python = ["3.10", "3.11", "3.12", "3.13"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "W"]
ignore = ["D100", "D104"]
